[TECH-272] Rename misspelt traefik
branch: feature/TECH-272-upgrade-monorepo
