+++ src/mpyl/steps/deploy/k8s/chart.py
@@ -17,7 +17,7 @@ from . import get_namespace
 from .resources import CustomResourceDefinition, to_dict  # pylint: disable = no-name-in-module
 from .resources.sealed_secret import V1SealedSecret
 from .resources.spark import to_spark_body, get_spark_config_map_data, V1SparkApplication
-from .resources.treafik import V1AlphaIngressRoute, V1AlphaMiddleware, \
+from .resources.traefik import V1AlphaIngressRoute, V1AlphaMiddleware, \
     HostWrapper  # pylint: disable = no-name-in-module
 from ...models import Input, ArtifactType
 from ....project import Project, KeyValueProperty, Probe, Deployment, TargetProperty, Resources, Target, Kubernetes, \
